SearchFacility Diffusable name fix
