+++ app/services/search_facility/diffusable.rb
@@ -4,36 +4,46 @@ module SearchFacility
       return blank_query if @query.blank?
       siren = FormatSiret.siren_from_query(@query[0..8])
       begin
-        data_company = ApiEntreprise::Entreprise::Base.new(siren).call[:entreprise]
-        data_siege = ApiEntreprise::Etablissement::Base.new(data_company["siret_siege_social"]).call[:etablissement]
-        formatted_items(data_siege, data_company)
-      rescue ApiEntreprise::ApiEntrepriseError => e
+        response = ApiInsee::SiretsBySiren::Base.new(siren).call
+        items = response[:etablissements_ouverts].map do |entreprise_params|
+          next if entreprise_params.blank?
+          ApiConsumption::Models::FacilityAutocomplete::ApiInsee.new({
+            nombre_etablissements_ouverts: response[:nombre_etablissements_ouverts],
+            un_seul_etablissement: true,
+            etablissement: entreprise_params.except('uniteLegale'),
+            entreprise: entreprise_params['uniteLegale']
+          })
+        end
+        return { items: items, error: nil }
+      # fallback si l'API Insee est en carafe
+      rescue ApiInsee::UnavailableApiError => e
+        from_full_text
+      rescue ApiInsee::ApiInseeError => e
         message = e.message.truncate(1000) # Avoid overflowing the cookie_store with alert messages.
         return { items: [], error: message }
       end
-    end
+  end
 
     def from_siret
       return blank_query if @query.blank?
       siret = FormatSiret.siret_from_query(@query[0..13])
-      siren = FormatSiret.siren_from_query(@query[0..8])
       begin
-        data_company = ApiEntreprise::Entreprise::Base.new(siren).call[:entreprise]
-        data_facility = ApiEntreprise::Etablissement::Base.new(siret).call[:etablissement]
-        formatted_items(data_facility, data_company)
-      rescue ApiEntreprise::ApiEntrepriseError => e
+        response = ApiInsee::Siret::Base.new(siret).call
+        items = response[:etablissements].map do |entreprise_params|
+          next if entreprise_params.blank?
+          ApiConsumption::Models::FacilityAutocomplete::ApiInsee.new({
+            nombre_etablissements_ouverts: response[:nombre_etablissements_ouverts],
+            un_seul_etablissement: true,
+            etablissement: entreprise_params.except('uniteLegale'),
+            entreprise: entreprise_params['uniteLegale']
+          })
+        end
+        return { items: items, error: nil }
+      # pas de fallback pour le moment, on trouve pas d'API "equivalente"
+      rescue ApiInsee::UnavailableApiError, ApiInsee::ApiInseeError => e
         message = e.message.truncate(1000) # Avoid overflowing the cookie_store with alert messages.
         return { items: [], error: message }
       end
-    end
-
-    private
-
-    def formatted_items(facility, company)
-      return {
-        items: [ApiConsumption::Models::FacilityAutocomplete::ApiEntreprise.new({ etablissement: facility, entreprise: company })],
-        error: nil
-      }
     end
   end
 end
